Only send Minecraft messages from within the last 2 seconds
This is necessary because upon bot startup, if the Exaroton server is already
online, it sends the client all the previous console lines, rather than just
those that were created after registering the client. The bot assumes that the
logs contain datetime information, rather than just time, which is achieved
using a custom log4j configuration file on the Exaroton server.